+++ server/tests/client_future_tests.rs
@@ -172,13 +172,15 @@ fn create_sig0_ready_client(io_loop: &Core) -> (BasicClientHandle, domain::Name)
   let signer = Signer::new(Algorithm::RSASHA256,
                            key,
                            domain::Name::with_labels(vec!["trusted".to_string(), "example".to_string(), "com".to_string()]),
-                           Duration::max_value());
+                           Duration::max_value(),
+                           true,
+                           true);
 
   // insert the KEY for the trusted.example.com
   let mut auth_key = Record::with(domain::Name::with_labels(vec!["trusted".to_string(), "example".to_string(), "com".to_string()]),
                                   RecordType::KEY,
                                   Duration::minutes(5).num_seconds() as u32);
-  auth_key.rdata(RData::KEY(DNSKEY::new(false, false, false, signer.get_algorithm(), signer.get_key().to_vec().expect("to_vec failed"))));
+  auth_key.rdata(RData::KEY(DNSKEY::new(false, false, false, signer.get_algorithm(), signer.get_key().to_public_bytes().expect("to_vec failed"))));
   authority.upsert(auth_key, 0);
 
   // setup the catalog